requirements.txt
azure-functions
python-dotenv
stripe
